[MINOR][PYTHON][SDP] Move Auto CDC flow tests into a dedicated test file
Extract the `create_auto_cdc_flow` test cases that were added to
`test_graph_element_registry.py` into a new `test_auto_cdc_flow.py`
so future Auto CDC tests have a natural home.

- Move the 8 Auto CDC flow tests into `AutoCdcFlowTest` in the new file,
  guarded by `should_test_connect` per the pipelines test convention.
- Keep the `create_auto_cdc_flow` outside-registry assertion in
  `test_definition_without_graph_element_registry`.
- Register the new test module in `dev/sparktestsupport/modules.py`.
